Move etag/mtime logic to more generic HTTP crawler
